==> ML_part/EOD/1_Data_cleaning.ipynb <==
Import the required libraries

from jupyter_core.paths import jupyter_path
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import seaborn as sns
import os
import glob
import matplotlib.pyplot as plt
from datetime import date
from datetime import datetime
from datetime import timedelta

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import TimeSeriesSplit
from sklearn import metrics
from sklearn.metrics import mean_absolute_error as mae
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.metrics import precision_score
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import KFold
import lightgbm as lgb # Our ML library

#Cross validation libraries
from sklearn.model_selection import TimeSeriesSplit

import joblib

Read all CSVs with stocks data and append to one big file

#os.chdir("/Users/olegkazanskyi/Documents/GitHub/Trading/CSVs")
os.chdir("C:/Users/oleg.kazanskyi/OneDrive - Danaher/Documents/Trading/SP500_CSVs_01032023")
filepaths = [f for f in os.listdir("./") if f.endswith('.csv')]
df = pd.DataFrame()
for i in filepaths:
    iterate_df = pd.DataFrame()
    iterate_df = pd.read_csv(i, encoding= 'unicode_escape')
    iterate_df["stock"] = i[:-4]
    df = pd.concat([df,iterate_df])
#df = pd.concat(map(pd.read_csv, filepaths))

df = df[df.close.notna()]
os.chdir("C:/Users/oleg.kazanskyi/OneDrive - Danaher/Documents/Trading/ML_part/EOD")
df.to_csv("THE_FINAL_DATASET_2023.csv")

record the dataframe to speed up the future reading process

os.chdir("C:/Users/oleg.kazanskyi/Personal-oleg.kazanskyi/Trading Python/")
df.to_csv("THE_FINAL_DATASET_2023.csv")

os.chdir("C:/Users/oleg.kazanskyi/Personal-oleg.kazanskyi/Trading Python/ML_part/EOD")
df= pd.read_csv("THE_FINAL_DATASET_2023.csv")
#df.drop(["Unnamed: 0","dataCode","name","description","statementType","units","Unnamed: 0.2", "Unnamed: 0.1"], axis = 1, inplace = True)
df = df[df.close.notna()]

#df['YoY_DY'] = 100*df['DY'].pct_change(periods = -365)

df.shape

df.duplicated().sum()

Set up Date column as an index columns

df["date"] = pd.to_datetime(df["date"])

Let's check how much empty values we have by column

zeroes = df.isnull().sum()
print(zeroes[zeroes>0])
del zeroes

### Dealing with "EPS" column

There are many null values in the "current_ratio" column. Let's see the stocks where it happens and then decide what to do

df[df.EPS_surprise.isnull()].stock.unique()

After checking the original files, it's apparently the null values comes from the older data.

Before 2017 our API did not provide us info about earinings. 

We can drop the reocrds with 'nan' values

we will remove only 10k records 

df = df[df.EPS_surprise.notnull()]

Let's check the dataset again

def check_null_cols(df):
    zeroes = df.isnull().sum()
    print(zeroes[zeroes>0])
    
check_null_cols(df)   

### Dealing with the null values in the "dividends" column

About 30% of data is affected by null values in the dividends columns.
Let's check how many companies are in the list

print("Number of companies with zero values in dividends",len(df[df['DY'].isnull()].stock.unique()))

This companies do not pay dividends, we can replace payments to "0"

df['DY'].fillna(0, inplace=True)
df['YoY_DPR'].fillna(0, inplace=True)
df['DPR'].fillna(0, inplace=True)
df['YoY_DY'].fillna(0, inplace=True)                   

Let's mark the companies with no dividends with the categorical columnm

df.loc[df['DY'] == 0,"Pays_Divds"] = 1
df.loc[df['DY'] != 0,"Pays_Divds"] = 0

temporary = df[df['DY'] == 0].copy()

print("Doesn't pay dividends")
for column in temporary.columns:
    if temporary[column].nunique() < 20:
        print(column, temporary[column].nunique())
        
temporary = df[df['DY'] != 0].copy()

print("\nPay dividends")
for column in temporary.columns:
    if temporary[column].nunique() < 20:
        print(column, temporary[column].nunique())
    

### Dealing with YoY

The YoY variables with null values are caused by YoY calculation of the rows without historical data.
Let's drop these

df = df[df.YoY_CR.notnull()]

Let's check the dataset again

check_null_cols(df)   

df.shape

### Dealing with debt ratio

Let's check the companies that contain null values

df[df['DE'].isnull()].stock.unique()

It seems these companies had zero debt for some quarters.

Let's replace with "0"

df['DE'].fillna(0, inplace=True)
df['LTDE'].fillna(0, inplace=True)

check_null_cols(df)   

### Dealing with VIX

Let's remove null values for VIX as these are related to the historical calculations

df = df[df.VIX_MoM.notnull()]

check_null_cols(df) 

### Dealing with Accounts Payable

df[df['Acc_Rec_Pay_Ration'].isnull()].stock.unique()

It seems these companies had zero Accounts receivables for some quarters

df['Acc_Rec_Pay_Ration'].fillna(0, inplace=True)

### Dealing with PEG

df[df['PEG_Forward'].isnull()].stock.unique()

We can drop the rows as the effect would be unsignificant

df = df[df.PEG_Forward.notnull()]

df[df['PEG_Backwards'].isnull()].stock.unique()

df = df[df.PEG_Backwards.notnull()]

df = df[df.EPS_1Y_exp_Change.notnull()]

check_null_cols(df) 

### Dealing with Industry

df[df['sector'].isnull()].stock.unique()

Let's fill the sector and industry values for the GEN company

df.loc[df['stock'] == 'GEN','sector'] = 'Information Technology'
df.loc[df['stock'] == 'GEN','industry'] = 'Software & Services'

check_null_cols(df) 

df = df[df.YoY_AR_Ration.notnull()]

### Working with data types

Now we can keep only the necessery columns

categoric_columns = df.select_dtypes(include='object').columns
for col in categoric_columns:
    if col == "stock":
        continue
    print(f"column {col}, data: \n {df[col].unique()}")

#Number of stocks per industry
df[['sector','industry','stock']].groupby('industry').stock.nunique()

#Number of stocks per sector
df[['sector','industry','stock']].groupby('sector').stock.nunique()

We can drop "Stocks" and "Industry" columns as there are too many unique values that block us from generalizing the data. 

#df.drop(["industry", "stock"], axis = 1, inplace = True)
#df.drop(["industry"], axis = 1, inplace = True)

Let's also check if there are any infinite numbers that can cause same trublesas nan

df.replace([np.inf, -np.inf], np.nan, inplace = True)

check_null_cols(df) 

df = df[df.EPS_1Y_exp_Change.notnull() & df.EPS_YoY_Growth.notnull() & df.EPS_QoQ_frcst_diff.notnull() & df.EPS_1Y_exp_Change_QoQ.notnull()]

check_null_cols(df)

#df = df[df['Fed_Balance_YoY'].notnull()]

### Data Cleaning is Over!

## Trimming the data to avoid overfitting

 Our dataset have the range of dates that are very close to each other. 

The changes in trading value are rarely very different between today and yesterday. 

There are exemptions but mostly these are connected with big surprises, and in any case we would notice the change even if we track every other day or every 4th day.

It means that removing part of the dataset should help us in building more generalized model since we will be looking for trend but not for matching values.


Let's do this, let's remove 4/5 of the dataset.

That means we will keep only one day of data per week. It will help us to generalize the dataset

print("Old dataframe shape:", df.shape)
df_compact = df.iloc[::5, :]
print("New dataframe shape:", df_compact.shape)

The final dataset is 131K rows long with 53 variables and 6 targets(price after 15 days, 30 days, 60 days, 90 days, 120 days and 150 days)

We can remove the bigger dataset, but let's save it to the file before doing so.

os.chdir("C:/Users/oleg.kazanskyi/OneDrive - Danaher/Documents/Trading/ML_Part/EOD")
df.to_csv("full_cleaned_dataframe_2023.csv", index = False, header = True)
del df

Lets' save the shorter version of the dataframe as well so we can get it faster when required

os.chdir("C:/Users/oleg.kazanskyi/Personal-oleg.kazanskyi/Trading Python/ML_Part/EOD")
df_compact.to_csv("shorter_cleaned_dataframe_2023.csv", index = False, header = True)
