==> ML_part/EOD/6_Saving_Model.ipynb <==
# Final Word

On average with every winning bet we can get about 5.8% of stock value. If it is true, it's an amazing result for 14 days trade.

We see that some bets lead to loss of  15.5%, at the same time max gain of the win bet is 33%.

Overall the results looks promising, and the level of precision is way higher than I expected after testing the model on validation.

I really would like to go live with the model.

import joblib

### Save the model as joblib

#Saving the label encoder
filename = 'C:\\Users\\oleg.kazanskyi\\OneDrive - Danaher\\Documents\\Trading Python\\Models\\label_encoder_012023.joblib'
joblib.dump(le, filename)

#Saving the models
for i, period in enumerate(periods):
    folder = 'C:\\Users\\oleg.kazanskyi\\OneDrive - Danaher\\Documents\\Trading Python\\Models\\'
    full_path = os.path.join(folder,period+"_model.joblib")
    joblib.dump(list_of_classification_models[i], full_path)
